+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
+* Check for ScopedTypeVariables being enabled when using withStatic.
+
 ### Changed
 
 ## [0.3.5] - 2018-02-01